fix(file-utils): fix could not find prefix for file, adding find pref… (#191)
* fix(file-utils): fix could not find prefix for file, adding find prefix again function and automatically adding to the system path

This code will first try to find the prefix in the current sys.path. If it fails, it will add the current working directory to sys.path and try to find the prefix again. If it still fails, it will raise a ValueError with instructions on how to add the current working directory to sys.path manually.

* ci(codecov.yml): add config for codecov to ignore certain folders outside the main library